get rid of dox
