+++ package.json
@@ -78,7 +78,6 @@
     "@cucumber/cucumber": "^7.0.0",
     "del": "^5.1.0",
     "depcheck": "^1.0.0",
-    "dox": "^0.9.0",
     "ejs": "^3.0.2",
     "eslint": "^7.10.0",
     "husky": "^4.2.3",
